Downgrade CF CLI version used in CI
Our CI foundation is currently using TAS 2.4, which is an older version. The
CF CLI has continued to release new versions, and now the newest releases are
no longer compatible with our older foundation. We have work scheduled to
update our remote foundation, but until then we need to continue using a CF
CLI version that is compatible with our deployed CF API. This commit hardcodes
the CF CLI version that we use in our CI pipelines for this project, for now.